[#5050] Refactor common functionality into LineBreakCorrector
