+++ lib/rubocop.rb
@@ -146,6 +146,7 @@ require_relative 'rubocop/cop/mixin/unused_argument'
 require_relative 'rubocop/cop/correctors/alignment_corrector'
 require_relative 'rubocop/cop/correctors/condition_corrector'
 require_relative 'rubocop/cop/correctors/empty_line_corrector'
+require_relative 'rubocop/cop/correctors/line_break_corrector'
 require_relative 'rubocop/cop/correctors/multiline_literal_brace_corrector'
 require_relative 'rubocop/cop/correctors/ordered_gem_corrector'
 require_relative 'rubocop/cop/correctors/parentheses_corrector'
